Add `.editorconfig` for convenience (#224)
The `Convenientions` section in `CONTRIBUTING` emphasizes an
indentation of 4 spaces which is not always the default in the
developer's environment. The EditorConfig project aims to maintain
consistent styles across various IDEs and editors. Some editors
natively support EditorConfig [1] and for some editors a plugin to
support EditorConfig can be downloaded [2].

[1] https://editorconfig.org/#pre-installed
[2] https://editorconfig.org/#download

Signed-off-by: Nicolas Bock <nicolas.bock@canonical.com>