libnetplan: Properly mark the lib/bin interface
This patch cleanly marks which functions/objects are part of the public
library API, and which are meant to be consumed by our own binaries.

In order to reduce as much as possible the ABI dependencies between our
binaries and the library, I move the various generator modules into the
library as it make sense for them to access directly the objects. On the
other hand, both generate.c and dbus.c should be relatively trivial to
change to use getters and setters instead of direct structure access.

Those changes will be the object of later patches.

V3:
  * Makefile: consolidate the pkg-config calls and add back the LDFLAGS
    variable
  * Normalize the touched function declarations to always have the
    symbol name at the start of the line.