lib: helper functions to reset some internal data structures
Those functions do a full in-depth cleaning and reset each value to its
default.

I chose to put them in a separate source file as I feel we're getting
away from the notion of parsing and much more into the object management
after its creation. In all honesty, if the netplan_netdef_new function
didn't reference a static data structure, I would have moved it into the
new module as well (and probably will in my later API-breaking patchset,
but that's a story for another time).