lib: clarify the ownership of the netdef objects
With this patch, we now decide arbitrarily that the owner of the netdef
objects is the netdef_ordered list, and all other pointers are weak
references. Knowing this, we can now easily clean up all netdefs.

Since _ordered is the owner of the global netdefs, it doesn't make sense
to change the cur_netdef reference before the new object is in the
global array, which is why we do the netdef initialization on a
local-only object, which we hand over only after it has been registered
in _ordered.