NetDefinition: ownership and cleanup (FR-786) (#228)
This PR is mostly about cleaning up the NetplanNetDefinition objects, and when to do it. There are a couple of precursor commits that either made the transition easier or fixed some bugs that were uncovered when actually trying to clean up things ;).

There are no API or ABI change, although there's one observable change with the dhcp_identifier which now defaults to NULL instead of an owned string "duid". The rationales are in the commit that makes the change.

As usual, each commit log contains some context information related to each change, and the commits are split up with the explicit intent of making it easier to review, one commit at a time.

COMMITS:
* lib: parse-nm: clarify ownership of the backend-settings strings
Some of the strings that were stored in the generated backend settings
had unclear ownership, leading to either leaks, double-free, or dangling
pointers. This patch forces the duplication in order to simplify
handling afterwards, and documents some assumptions that aren't shown
through the typesystem (I'm looking at you, GData *).

* lib: use NULL for the dhcp_identifier default case
This centralizes the place where we know about "duid", and makes my life
easier when writing reset/freeing code for netdef as I don't have to
reallocate a string after I've cleared them all.

* lib: helper functions to reset some internal data structures
Those functions do a full in-depth cleaning and reset each value to its
default.
I chose to put them in a separate source file as I feel we're getting
away from the notion of parsing and much more into the object management
after its creation. In all honesty, if the netplan_netdef_new function
didn't reference a static data structure, I would have moved it into the
new module as well (and probably will in my later API-breaking patchset,
but that's a story for another time).

* lib: clarify the ownership of the netdef objects
With this patch, we now decide arbitrarily that the owner of the netdef
objects is the netdef_ordered list, and all other pointers are weak
references. Knowing this, we can now easily clean up all netdefs.
Since _ordered is the owner of the global netdefs, it doesn't make sense
to change the cur_netdef reference before the new object is in the
global array, which is why we do the netdef initialization on a
local-only object, which we hand over only after it has been registered
in _ordered.