tests: normalize multiple keys on a single dict
The YAML normalizer has some special cases for some given keys. However,
the code block where those are dealt with used `elif` everywhere, which
meant that if a single dict had multiple keys needing to be dealt with,
only the first one would be fixed, and the rest would continue on, still
broken.

This bug is latent in the current codebase, but it showed up in a
patchset I'm working on, where by the game of butterfly effects the
rendered configuration would insist on having proper renderer for each
netdef. This key needs normalizing, but in the test cases where the
netdefs have other problematic properties, for instance the GSM config
in `test_cdma_config`, those pesky renderer lines would still show up
and screw up the test.

V2:
* Beat pycheckstyle into submission in a much more sensible fashion,
  thanks to @slyon