lib: fix error path on ip_rules parsing
The previous code flow had several issues. First, if the caller provided
NULL as the GError, then errors in individual rule objects were ignored
and the object simply dropped unless the from/to check failed.

Second, the parsed object was added to the netdef before all checks were
finished, which meant that in case of an error, the netdef was left in
an invalid state.

This patch uses a return-value-based error flow instead, and moves the
cur_netdef bits to the end of the process when we know the object is
valid. Also, we actually do some cleanup on the error path to prevent
the buggy rule leaking out into the world. This is only superficial, so
there will be some leakage from the subobjects, but this will be handled
in a later patchset.

V2: remove braces on single-line 'if' bodies