lib: rework the wireguard peer error flow
This new version checks the validity of a given peer as early as possible,
before impacting the netdef, allowing us to keep the netdef without provably
erroneous data.

While we're in there, we inserted some minimal cleanup, although a more
thorough leak plugging will be the object of a later patch.