lib: rework the AP parsing error flow
This new version checks the validity of a given AP as early as possible,
starting with SSID conflict checks before even parsing the body, and
only impacts the netdef at the end of the loop body.

This allows us to keep the netdef without provably erroneous data, and
also consolidates the error path, where we can easily add some cleanup
code, which we started to do. Said cleanup will be the object of a
further patchset, though.