parse-nm: Handle missing gateway in keyfile routes, keep dns-search fallback
NM assumes a route to use the unspecified address as the gateway
(via = "0.0.0.0"/"::") if none is specified in the keyfile.
E.g. the route is only valid on the local network:
"ip route add NETWORK dev DEVICE [metric METRIC]"

netplan cannot differentiate between ipv4.dns-search and ipv6.dns-search
so keep it in the passthrough/fallback list as an override.