lib: networkd: normalize the API with proper error handling
Note that we duplicated a function, as I'd like to remove it from the
API seeing as it's really not netplan-specific, but we need to keep it
in the ABI.

The new API has the ability to report on errors, which allows us to
remove all the calls to exit() within networkd.c. Those have nothing to
do in library code.

We also followed the current convention of using the return value to
signal errors, forcing us to use an out-parameter for the
"has_been_written" return value.