parse-nm: Handle missing gateway in keyfile routes, keep dns-search fallback (#238)
* parse-nm: Handle missing gateway in keyfile routes, keep dns-search fallback

NM assumes a route to use the unspecified address as the gateway
(via = "0.0.0.0"/"::") if none is specified in the keyfile.
E.g. the route is only valid on the local network:
"ip route add NETWORK dev DEVICE [metric METRIC]"

netplan cannot differentiate between ipv4.dns-search and ipv6.dns-search
so keep it in the passthrough/fallback list as an override.

* parse-nm: Fix typos

Co-authored-by: Simon Chopin <simon.chopin@canonical.com>

* nm:parse-nm: Hide NM's automatic route-scope detection

NetworkManager automatically detects a route's scope, depending on
destination IP ("to") and gateway ("via"). If no gateway is specified
(e.g. the unspecified address "0.0.0.0"/"::" in keyfile) it will assume
a "link"/"host" scope, otherwise it will assume a "global" scope.

* nm: Do not set missing gateway as unspecified address

Also avoid the deprecated trailing comma notation at the same time.

Co-authored-by: Simon Chopin <simon.chopin@canonical.com>