parser: fix some error handling code (#234)
* lib: fix error path on ip_rules parsing

The previous code flow had several issues. First, if the caller provided
NULL as the GError, then errors in individual rule objects were ignored
and the object simply dropped unless the from/to check failed.

Second, the parsed object was added to the netdef before all checks were
finished, which meant that in case of an error, the netdef was left in
an invalid state.

This patch uses a return-value-based error flow instead, and moves the
cur_netdef bits to the end of the process when we know the object is
valid. Also, we actually do some cleanup on the error path to prevent
the buggy rule leaking out into the world. This is only superficial, so
there will be some leakage from the subobjects, but this will be handled
in a later patchset.

V2: remove braces on single-line 'if' bodies

* lib: rework the AP parsing error flow

This new version checks the validity of a given AP as early as possible,
starting with SSID conflict checks before even parsing the body, and
only impacts the netdef at the end of the loop body.

This allows us to keep the netdef without provably erroneous data, and
also consolidates the error path, where we can easily add some cleanup
code, which we started to do. Said cleanup will be the object of a
further patchset, though.

* lib: rework the wireguard peer error flow

This new version checks the validity of a given peer as early as possible,
before impacting the netdef, allowing us to keep the netdef without provably
erroneous data.

While we're in there, we inserted some minimal cleanup, although a more
thorough leak plugging will be the object of a later patch.

* lib: netplan_finish_parse: behave consistently when error == NULL

The previous version would simply disregard any failed checks if the
caller provided error as NULL. This is problematic as the caller could
simply not be interested in the detail of the errors, but would still
like to know if they have a functional configuration or not...

* Fix coding style

Co-authored-by: Lukas Märdian <slyon@ubuntu.com>