libnetplan: expose coherent generator APIs (#239)
New APIs for the generator APIs of libnetplan, with proper error handling and support for the NetplanState feature.

API and ABI compatibility is retained for now. Depends on #232 and #234, as usual see the commits after the merge.

COMMITS:
* lib: networkd: normalize the API with proper error handling

Note that we duplicated a function, as I'd like to remove it from the
API seeing as it's really not netplan-specific, but we need to keep it
in the ABI.

The new API has the ability to report on errors, which allows us to
remove all the calls to exit() within networkd.c. Those have nothing to
do in library code.

We also followed the current convention of using the return value to
signal errors, forcing us to use an out-parameter for the
"has_been_written" return value.

* lib: nm: normalize the API with error handling

The NM generator directly called exit() when something went wrong, which
is not ideal in library code!

We copied the API from the netword generator, hence the has_been_written
new capability.

* lib: openvswitch: normalize the API with proper error handling

The API is now the same as for the networkd and NM generators.

No more exit()ing in the middle of a library call!

* openvswitch: formatting cleanup
* test_routing: fix coverage
* nm: formatting cleanup
* generate: do not drop comments
* networkd: formatting & comments

Co-authored-by: Lukas Märdian <slyon@ubuntu.com>