lib:netplan: output all the dirty fields in the YAML dump
If the user touched those fields, we want to retranscribe the value even
if it matches the default!

Squashed in this commit is a test fix for the on-link field.
on-link is a boolean, and its output format was wrong. We want literals
as much as possible! The behaviour change is entangled with the dirty
changes, so we change the test here.

V2: Change the macro names to remove the DIRTY_ bit and make the default
values more explicit