netplan: start introducing high-level classes in the netdef iterator
We're using the iterator as the introduction point of the transition,
but the idea is to use these classes as the main way to interact with
the C library.

Note the temporary GlobalState class that has been introduced as a
temporary measure. It should be gone by the end of the patchset.
:)

V2:
* rename struct devtype_def_iter to netdef_pertype_iter