lib: ensure consistency in OVS port input/output
The previous implementation relied on the order of traversal of a
hashtable, which isn't exactly reliable. It wasn't problematic because
the semantics of the port pair are symetrical, but since we're planing
on using the YAML netplan generator backend as a passthrough for
`netplan set`, we want to preserve the user's input as pristinely as
possible.

The new implementation uses the ordered list of netdefs to determine the
order in which netdefs are processed, which is both deterministic and
sensible, since it respects the order in which the user fed us the info
(modulo the grouping by devtype).