lib:util: new helper function to dump a yaml subtree
This API is also using file descriptors for flexibility. The idea is to
be able to copy only part of a YAML document based on a "prefix", that
is subsequent keys into YAML maps, separated by '\t' characters.

V2: add some comments and use better names for the static helpers