netplan:cli:get: remove the PyYAML dependency
This patch rewrites the entire command to use our internal YAML helpers
instead, notably the APIs introduced in the few previous commits.

This also changes the tests to parse the output YAML instead of checking
for a *specific* output, which makes them less susceptible of breaking
when there's a small formatting change.

V2:
* add a test for the networkINVALID query
* remove a spurious change in tests/generator/base.py