get/set: ignore empty YAML hints and delete files on "set network=null" (LP: #1946957) (#246)
This PR contains several improvements wrt. the handling of YAML hint files, as used by the netplan set CLI.

If a --origin-hint is specified, that already exists as an empty file, that file is ignored.
If netplan set network=null [--origin-hint FILE] is executed, a special handler makes sure that either the specified origin hint file is remove, or all files in etc/netplan/*.yaml are cleared otherwise.

COMMITS:
* test:get-set: detect unexpected output in CLI
* cli:set: ignore empty hint files
* cli:set: special case for 'network=null' to clear YAML files
* cli:set: improve logging
* cli:set: do not crash if we try to unset something that does not exist (LP: #1946957)
* cli:set: fix another case of writing an empty origin-hint file (containing whitespace)