YAML consolidation prelude: OVS handling (FR-702) (#249)
* lib: ensure consistency in OVS port input/output

The previous implementation relied on the order of traversal of a
hashtable, which isn't exactly reliable. It wasn't problematic because
the semantics of the port pair are symetrical, but since we're planing
on using the YAML netplan generator backend as a passthrough for
`netplan set`, we want to preserve the user's input as pristinely as
possible.

The new implementation uses the ordered list of netdefs to determine the
order in which netdefs are processed, which is both deterministic and
sensible, since it respects the order in which the user fed us the info
(modulo the grouping by devtype).

* netplan:names: give a proper name to the OVS port pseudo-deftype

Later in the patchset we expose this data to Python, and are using the
names rather than the integers to represent the netdef types. As a
result, we need a proper name for this type to distinguish it from the
default value.

* lib:ovs: change the ovs-port name to mark it as "special"

Contrary to all other def types, this one is "virtual", as it doesn't have a YAML key matching it. Prefixing it with an underscore should make it obvious that *something* is different there.

Co-authored-by: Lukas Märdan <slyon@ubuntu.com>