YAML consolidation prelude: High-level libnetplan bindings (FR-702) (#250)
This PR introduces some high-level Python bindings. This is mostly a skeleton that will be populated with the needed elements in later PRs, which will remain in draft state until this one is merged.

COMMITS:
* netplan: split off the bindings into their own module
The utils module was getting kind of crowded.

* netplan: start introducing high-level classes in the netdef iterator
We're using the iterator as the introduction point of the transition,
but the idea is to use these classes as the main way to interact with
the C library.

Note the temporary GlobalState class that has been introduced as a
temporary measure. It should be gone by the end of the patchset.
:)

V2:
- rename struct devtype_def_iter to netdef_pertype_iter

* lib: types: add explicit non-null assertions to public-facing getters

Co-authored-by: Lukas Märdian <slyon@ubuntu.com>