YAML consolidation prelude: new YAML dump APIs (FR-702) (#251)
Introduce a couple of APIs that are needed for the various YAML consolidation efforts. Note that this is technically an API break, but the function that is removed has never been part of a release, so this should be OK.

COMMITS:
* lib:netplan: rework API to dump the state into a file descriptor
This allows for more flexibility than using filenames. The
implementation is split in two functions to allow for future variants,
i.e. updating an existing netplan configuration tree.

This patch is technically an API break but the removed API was never
released.

V2:
- use a helper function netplan_state_has_nondefault_globals()
- fix the write_netplan_conf_full behavior on empty state to not create
  an empty file
- DRASTICALLY simplify the implementation of netplan_state_dump_yaml(),
  fixing a leak by not using a new list in the first place

* lib:yaml-helpers: add helpers to produce yaml NULL values
* lib:util: new helper function to dump a yaml subtree
This API is also using file descriptors for flexibility. The idea is to
be able to copy only part of a YAML document based on a "prefix", that
is subsequent keys into YAML maps, separated by '\t' characters.

V2: add some comments and use better names for the static helpers