parse: fix the handle_match_driver type
The YAML_NO_NODE type marker means that the callback called is via the
`.variable` union field, not `.generic`. Since this is a union type, and
both callbacks have the same signature for now, it works, but future
evolutions of the code will make those signatures evolve independently,
leading to hard-to-debug issues.