Drop unused legacy/compat API (#262)
Libnetplan at out latest version 0.103 did not expose several symbols that we carry along (and would newly introduce as "legacy"/"compat" symbols in 0.104). This is due to the fact that the Makefile changed and those functions had been linked directly into the generate binary before, whereas they are now part of libnetplan.so. As those symbols have not been part of the library before, we should not introduce them now as legacy symbols, as no consumer can be using it from the old library.

COMMITS:
* openvswitch: make use of new netplan_netdef_write_network_file() API
* abi_compat: drop legacy API that has never been part of the library